LayoutTests/platform/glib/imported/w3c/web-platform-tests/mathml/relations/css-styling/display-2-expected.txt


FAIL block display assert_approx_equals: block size expected 49 +/- 1 but got 45
FAIL block display with contrained width assert_approx_equals: block size expected 78 +/- 1 but got 70
FAIL list display inside display block assert_approx_equals: block size expected 194 +/- 1 but got 78
FAIL inline display assert_approx_equals: block size expected 29 +/- 1 but got 41
FAIL inline-block display assert_approx_equals: block size expected 29 +/- 1 but got 41
FAIL table display (math) assert_approx_equals: block size expected 78 +/- 1 but got 95
FAIL table display (mrow) assert_approx_equals: block size expected 78 +/- 1 but got 95
FAIL inline-table display (math) assert_approx_equals: block size expected 58 +/- 1 but got 75
FAIL inline-table display (mrow) assert_approx_equals: block size expected 78 +/- 1 but got 95
FAIL flexbox display (math) assert_approx_equals: block size expected 95 +/- 1 but got 45
FAIL flexbox display (mrow) assert_approx_equals: block size expected 95 +/- 1 but got 45
FAIL grid display (math) assert_approx_equals: block size expected 99 +/- 1 but got 45
FAIL grid display (mrow) assert_approx_equals: block size expected 99 +/- 1 but got 45
FAIL ruby display (math) assert_approx_equals: block size expected 78 +/- 1 but got 41
FAIL ruby display (mrow) assert_approx_equals: block size expected 78 +/- 1 but got 45
FAIL block display with column width (math) assert_approx_equals: block size expected 49 +/- 1 but got 45
FAIL block display with column width (mrow) assert_approx_equals: block size expected 49 +/- 1 but got 45
